use gorm as orm.
Not really happy with gorm. Sort of a black box, v2 is coming, has bugs, not many features really. Will check later if it is easy to use with postgres too, but if changes are needed to code, we need to change this. Possibly just go back to using standard library and a simple sql-query-builder.